pydantic-ai-agent/pyproject.toml
[project]
name = "pydantic-ai-agent"
version = "0.1.0"
description = "PydanticAI Agent - The Standard Agent Core for production-grade type-safe AI"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.1.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


